refactor(cli): move update-check policy
