+++ src/cli/update-check-entry.ts
